backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.4.0
numpy>=2.0.0
aiofiles>=23.2.1
